+++ example/windows/flutter/generated_plugins.cmake
@@ -6,7 +6,6 @@ list(APPEND FLUTTER_PLUGIN_LIST
   connectivity_plus
   firebase_auth
   firebase_core
-  flutter_secure_storage_windows
   local_auth_windows
   smart_auth
 )