+++ lib/src/data/sources/keep_data_source.dart
@@ -1,7 +1,7 @@
 import '../../models/auth.dart';
 import 'authorized_data_source.dart';
 
-class KeepDataSource<T extends Auth> extends BackupDataSourceImpl<T> {
+class KeepDataSource<T extends Auth> extends AuthorizedDataSourceImpl<T> {
   KeepDataSource({
     super.database,
   });